Fix v1 compilation (#628)
* auth-server: rfqt: Return relayer error on failed quote instead of panicking (#616)

When the relayer returns non-200 for a quote (e.g., 400 'order value too
small' or 204 'no match found'), handle_quote_and_assemble proceeded to
call sponsor_rfqt_quote_response which called ctx.response() — panicking
on the None response body. The panic killed the tokio task, dropping the
connection, which the load balancer surfaced as a 502 to the caller.

This only affected the malleable (default) RFQT path, which chains
quote + assemble internally. The direct path and non-RFQT quote endpoint
already had proper guards.

The fix checks quote_resp_ctx.is_success() after the quote step and
returns the relayer's error response directly, matching the pattern used
by every other post-request handler in the codebase.

* fix compilation against renegade v1 branch

---------

+++ funds-manager/funds-manager-server/src/custody_client/vaults.rs
@@ -125,7 +125,7 @@ impl CustodyClient {
             return Ok(None);
         };
 
-        let amount_f64 = available_f64 * 10_f64.powf(asset_onchain_data.decimals as f64);
+        let amount_f64 = available_f64 * 10_f64.powf(asset_onchain_data.decimals);
         let amount: u128 = amount_f64.floor() as u128;
 
         Ok(Some(TokenBalance { mint, amount }))
